backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
aiosqlite==0.20.0

# AI Agents
google-genai==0.3.0
anthropic==0.42.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# Exchange connectors
ccxt>=4.0.0
alpaca-py>=0.20.0

# Utilities
python-dotenv==1.0.1
pydantic-settings>=2.0.0
bcrypt==4.1.2
